Fixed bug with LogAnalyzer_Reachability.java
The program could not recognize custom leaks with multiple placeholders. To fix this, I limited the rawLeakText to only split() on the first occurance of "%d" in a leak. I also changed how the index was isolated from a leak so that it replaces non-index substrings with empty strings rather than using the split() function.