Fixed bug with DataLeak.java
Added two new methods that return the correct source/sink for a provided taint or scope OperatorType. When a source is set or get, it now checks to ensure that TAINTSINK and SCOPESINK will use TAINTSOURCE or SCOPESOURCE source leaks. When a sink is set or get, it also checks to ensure that TAINTSOURCE and SCOPESOURCE will use TAINTSINK or SCOPESINK sink leaks